Update Basic Attack Mechanic
Menambahkan mekanik elemental reaction, dan casting elemental magic projectile
