+++ Assets/Prefabs/BasicAttack/Pre_MagicProjectile.prefab
@@ -15,7 +15,7 @@ GameObject:
   - component: {fileID: 2709317378477034588}
   - component: {fileID: -4706335846477125676}
   m_Layer: 0
-  m_Name: MagicProjectile
+  m_Name: Pre_MagicProjectile
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -31,7 +31,7 @@ Transform:
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 0, y: 2.5, z: 2.2}
-  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_LocalScale: {x: 2, y: 2, z: 2}
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 0}
@@ -146,6 +146,5 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 11107556a749e73409ad0592a9e3056a, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  projectileSpeed: 10
   element: 0
   lifeTime: 1.5
