+++ Assets/Scenes/Demo/Demo_BasicAttack.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e2aafca7ceedf7b41aa5042d55b841ad
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
