+++ Assets/Scenes/Demo/Demo_BasicAttack/NavMesh-NavMesh Surface.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 758c12de60789864c894cfc8ff01e0aa
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 23800000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
