--- Assets/Scripts/BasicAttack/AttackController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bb21acf0f4cb1c445bdc72cdeb7751bd
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
